fix!: refactor Baggage to remove Noop*
