+++ propagator/ottrace/lib/opentelemetry/propagator/ottrace/text_map_propagator.rb
@@ -118,7 +118,7 @@ module OpenTelemetry
         end
 
         def baggage
-          OpenTelemetry.baggage
+          OpenTelemetry::Baggage
         end
 
         def inject_span_context(span_context:, carrier:, setter:)
