+++ propagator/ottrace/test/opentelemetry/propagator/ottrace/text_map_propagator_test.rb
@@ -57,7 +57,7 @@ describe OpenTelemetry::Propagator::OTTrace::TextMapPropagator do
   end
 
   let(:baggage) do
-    OpenTelemetry.baggage
+    OpenTelemetry::Baggage
   end
 
   let(:propagator) do
@@ -196,14 +196,6 @@ describe OpenTelemetry::Propagator::OTTrace::TextMapPropagator do
     end
 
     describe 'baggage handling' do
-      before do
-        OpenTelemetry.baggage = OpenTelemetry::Baggage::Manager.new
-      end
-
-      after do
-        OpenTelemetry.baggage = nil
-      end
-
       describe 'given valid baggage items' do
         it 'extracts baggage items' do
           carrier_with_baggage = carrier.merge(
@@ -212,8 +204,8 @@ describe OpenTelemetry::Propagator::OTTrace::TextMapPropagator do
           )
 
           context = propagator.extract(carrier_with_baggage, context: parent_context)
-          _(OpenTelemetry.baggage.value('foo', context: context)).must_equal('bar')
-          _(OpenTelemetry.baggage.value('bar', context: context)).must_equal('baz')
+          _(OpenTelemetry::Baggage.value('foo', context: context)).must_equal('bar')
+          _(OpenTelemetry::Baggage.value('bar', context: context)).must_equal('baz')
         end
       end
 
@@ -225,8 +217,8 @@ describe OpenTelemetry::Propagator::OTTrace::TextMapPropagator do
           )
 
           context = propagator.extract(carrier_with_baggage, context: parent_context)
-          _(OpenTelemetry.baggage.value('fθθ', context: context)).must_be_nil
-          _(OpenTelemetry.baggage.value('bar', context: context)).must_equal('baz')
+          _(OpenTelemetry::Baggage.value('fθθ', context: context)).must_be_nil
+          _(OpenTelemetry::Baggage.value('bar', context: context)).must_equal('baz')
         end
       end
 
@@ -238,8 +230,8 @@ describe OpenTelemetry::Propagator::OTTrace::TextMapPropagator do
           )
 
           context = propagator.extract(carrier_with_baggage, context: parent_context)
-          _(OpenTelemetry.baggage.value('foo', context: context)).must_be_nil
-          _(OpenTelemetry.baggage.value('bar', context: context)).must_equal('baz')
+          _(OpenTelemetry::Baggage.value('foo', context: context)).must_be_nil
+          _(OpenTelemetry::Baggage.value('bar', context: context)).must_equal('baz')
         end
       end
     end
@@ -326,14 +318,6 @@ describe OpenTelemetry::Propagator::OTTrace::TextMapPropagator do
     end
 
     describe 'baggage handling' do
-      before do
-        OpenTelemetry.baggage = OpenTelemetry::Baggage::Manager.new
-      end
-
-      after do
-        OpenTelemetry.baggage = nil
-      end
-
       describe 'given valid baggage items' do
         it 'injects baggage items' do
           context_with_baggage = baggage.build(context: context) do |builder|
